Исправлена неточность в расчетах цены
-Все поля денежных едениц (Price) теперь имеют тип decimal

+++ App/OrderModel.cs
@@ -6,5 +6,5 @@ public class OrderModel
     public string UserName { get; set; } = string.Empty;
     public OrderItemModel[] Items { get; set; } = [];
     public int TotalCount { get; set; }
-    public float TotalPrice { get; set; }
+    public decimal TotalPrice { get; set; }
 }
